+++ lib/features/transactions/data/models/transactions.dart
@@ -9,6 +9,7 @@ class Transactions extends Table {
   TextColumn get note => text().nullable()();
   late final occurredTime = dateTime().withDefault(currentDateAndTime)();
   late final creationTime = dateTime().withDefault(currentDateAndTime)();
+  late final updatedAt = dateTime().withDefault(currentDateAndTime)();
   TextColumn get categoryId => text().nullable().references(
     Categories,
     #id,