+++ test/helpers/mocks.dart
@@ -4,6 +4,7 @@ import 'package:stockflow/core/connectivity/domain/connectivity_repository.dart'
 import 'package:stockflow/core/connectivity/domain/usecases/watch_connectivity_usecase.dart';
 import 'package:stockflow/core/sync/data/datasources/supabase_sync_data_source.dart';
 import 'package:stockflow/core/sync/domain/repositories/sync_queue_repository.dart';
+import 'package:stockflow/features/transactions/data/transactions_data_source.dart';
 import 'package:stockflow/features/categories/domain/repositories/category_repository.dart';
 import 'package:stockflow/features/transactions/domain/repositories/transactions_repository.dart';
 import 'package:stockflow/features/categories/domain/usecases/add_category_usecase.dart';
@@ -52,3 +53,6 @@ class MockWatchConnectivityUseCase extends Mock
 class MockSyncQueueRepository extends Mock implements SyncQueueRepository {}
 
 class MockSupabaseDataSource extends Mock implements SupabaseSyncDataSource {}
+
+class MockTransactionsDataSource extends Mock
+    implements TransactionsDataSource {}